Move files out of the source folder
This way they don't end up compiled to the public folder
