+++ test/karma/TimeManager.test.js
@@ -1,5 +1,5 @@
 "use strict";
-var TimeManager = require("../../themes/microfolio/source/js/TimeManager")
+var TimeManager = require("../../themes/microfolio/components/js/TimeManager")
 
 var SECOND = 1000
 var MINUTE = 60 * SECOND
