+++ themes/microfolio/components/js/extensions/remove-element.js
