+++ themes/microfolio/source/css/index.styl
@@ -1,12 +1,12 @@
-@import "./settings"
+@import "../../components/css/settings"
 
 @font-face {
   font-family: "Fanwood";
-  src: url("webfonts/fanwood-webfont.eot");
-  src: url("webfonts/fanwood-webfont.eot?#iefix") format("embedded-opentype"),
-    url("webfonts/fanwood-webfont.woff") format("woff"),
-    url("webfonts/fanwood-webfont.ttf") format("truetype"),
-    url("webfonts/fanwood-webfont.svg#webfont") format("svg");
+  src: url("../webfonts/fanwood-webfont.eot");
+  src: url("../webfonts/fanwood-webfont.eot?#iefix") format("embedded-opentype"),
+    url("../webfonts/fanwood-webfont.woff") format("woff"),
+    url("../webfonts/fanwood-webfont.ttf") format("truetype"),
+    url("../webfonts/fanwood-webfont.svg#webfont") format("svg");
 }
 
 *
@@ -56,5 +56,5 @@ love
 buffer.noJS
   background-image url("/img/clouds.thumb.jpg")
 
-@require "./buffer"
-@require "./themes"
+@require "../../components/css/buffer"
+@require "../../components/css/themes"
